Separate create_book to another file
