+++ create_book.rb
@@ -1,0 +1,14 @@
+require './book'
+require './app'
+def create_book
+  print 'Title: '
+  title = gets.chomp
+
+  print 'Author: '
+  author = gets.chomp
+
+  book = Book.new(title, author)
+  @books.push(book)
+
+  puts 'Book created successfully'
+end
